# docs/explore/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SozoRock Place Intelligence API",
    "version": "2026-08-08",
    "description": "Versioned, non-clinical county evidence, planning-agent and collaboration contracts. Missing evidence remains missing; county evidence is never represented as ZIP-, city- or person-level evidence."
  },
  "servers": [
    {
      "url": "https://health.sozorockfoundation.org"
    }
  ],
...
  "paths": {
    "/api/evidence/v1/place-brief": {
      "get": {
        "tags": [
          "Evidence"
        ],
        "summary": "Return the pinned county evidence brief.",
        "deprecated": false,
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "county"
              ]
            }
          },
          {
            "name": "geoid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{5}$",
              "description": "Current Census county or county-equivalent GEOID."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ExplorePlaceBriefV1."
          },
          "400": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evidence/v1/voice/transcribe": {
      "post": {
        "tags": [
          "Voice"
        ],
        "summary": "Transcribe an intentionally recorded planning question without retaining raw audio.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "geoid",
                  "audio"
                ],
                "properties": {
                  "geoid": {
                    "type": "string",
                    "pattern": "^[0-9]{5}$",
                    "description": "Current Census county or county-equivalent GEOID."
                  },
                  "audio": {
                    "type": "string",
                    "contentEncoding": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reviewable transcript."
          },
          "400": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evidence/v1/workspaces/{workspaceId}": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "Load a membership-scoped county plan.",
        "security": [
          {
            "cognitoAccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace plan."
          },
          "403": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evidence/v1/heat-map": {
      "post": {
        "tags": [
          "Visuals"
        ],
        "summary": "Return compatible multi-county values and official boundaries.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountySetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cognitoAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Cognito access token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "CountySetRequest": {
        "type": "object",
        "required": [
          "geoids"
        ],
        "properties": {
          "geoids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 25,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "pattern": "^[0-9]{5}$",
              "description": "Current Census county or county-equivalent GEOID."
            }
          },
          "measureDefinitionId": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}